bugs fixed on user test
