+++ src/mocks/userMock.js
@@ -4,8 +4,8 @@ import bcrypt from 'bcrypt';
 export const mockedUser = {
   name: faker.name.findName(),
   email: faker.internet.email(),
-  password: `${faker.internet.password(8)}!`,
-  selectedCountry: faker.address.country,
+  password: `${faker.internet.password(8)}_9!`,
+  selectedCountry: `${faker.address.country()}`,
   birthDate: faker.date.past(),
   confirmedPassword() {
     return this.password;
